Add old datatable in Storybook
