--- src/App.vue
@@ -1,5 +1,0 @@
-<template>
-  <div>
-    Hello World
-  </div>
-</template>
