+++ src/Datatable/DatatableBlock.vue
@@ -1,34 +1,38 @@
+<!--eslint-disable-->
 <template>
-    <div class="datatable-block m-0 p-0" :style="style">
-        <div class="datatable-header-row m-0 p-0 border-bottom">
-            <DatatableHeaderCell
-                v-for="columnKey in Object.keys(columns)"
-                :tableConfig="tableConfig"
-                :key="'datatable-header-' + columnKey"
-                :columnId="columnKey"
-                :columnConfig="columns[columnKey]"
-                @changeOrder="changeOrder"
-            />
-        </div>
-        <div>
-            <div
-                class="datatable-data-row border-bottom m-0 p-0"
-                v-for="(line, index) in tableData"
-                v-if="isVisibleRow(line)"
-                :style="rowHeightStyle"
-            >
-                <DatatableCell
-                    v-for="columnKey in Object.keys(columns)"
-                    :tableConfig="tableConfig"
-                    :key="'datatable-cell-' + columnKey + '-' + index"
-                    :columnId="columnKey"
-                    :columnConfig="columns[columnKey]"
-                    :rowData="line"
-                    @toggleRowDisplay="toggleRowDisplay"
-                />
-            </div>
-        </div>
+  <div
+    class="datatable-block m-0 p-0"
+    :style="style"
+  >
+    <div class="datatable-header-row m-0 p-0 border-bottom">
+      <DatatableHeaderCell
+        v-for="columnKey in Object.keys(columns)"
+        :key="'datatable-header-' + columnKey"
+        :table-config="tableConfig"
+        :column-id="columnKey"
+        :column-config="columns[columnKey]"
+        @changeOrder="changeOrder"
+      />
     </div>
+    <div>
+      <div
+        v-for="(line, index) in tableData"
+        v-if="isVisibleRow(line)"
+        class="datatable-data-row border-bottom m-0 p-0"
+        :style="rowHeightStyle"
+      >
+        <DatatableCell
+          v-for="columnKey in Object.keys(columns)"
+          :key="'datatable-cell-' + columnKey + '-' + index"
+          :table-config="tableConfig"
+          :column-id="columnKey"
+          :column-config="columns[columnKey]"
+          :row-data="line"
+          @toggleRowDisplay="toggleRowDisplay"
+        />
+      </div>
+    </div>
+  </div>
 </template>
 
 <style lang="scss" scoped>
@@ -50,6 +54,9 @@
 </style>
 
 <script>
+/* eslint-disable */
+//TODO fix eslint
+
     import DatatableCell from "./DatatableCell";
     import DatatableHeaderCell from "./DatatableHeaderCell";
