+++ src/Datatable/DatatableCell.vue
@@ -1,25 +1,32 @@
+<!--eslint-disable-->
+
 <template>
+  <div
+    v-if="!columnConfig.hidden"
+    :style="style"
+    class="datatable-cell datatable-cell-data p-1"
+  >
     <div
-        v-if="!columnConfig.hidden"
-        :style="style"
-        class="datatable-cell datatable-cell-data p-1"
-    >
-        <div v-if="!isServiceColumn" v-html="content" />
-        <div v-if="isCollapseColumn && hasChildren">
-            <button class="btn btn-clear text-primary p-0 shadow-none"
-                v-if="!tableConfig.closedRows.has(rowData[tableConfig.rowId])"
-                @click="toggleRowDisplay(rowData[tableConfig.rowId])"
-            >
-              <b-icon icon="chevron-down" />
-            </button>
-            <button class="btn btn-clear text-primary p-0 shadow-none"
-                v-if="tableConfig.closedRows.has(rowData[tableConfig.rowId])"
-                @click="toggleRowDisplay(rowData[tableConfig.rowId])"
-            >
-              <b-icon icon="chevron-right" />
-            </button>
-        </div>
+      v-if="!isServiceColumn"
+      v-html="content"
+    />
+    <div v-if="isCollapseColumn && hasChildren">
+      <button
+        v-if="!tableConfig.closedRows.has(rowData[tableConfig.rowId])"
+        class="btn btn-clear text-primary p-0 shadow-none"
+        @click="toggleRowDisplay(rowData[tableConfig.rowId])"
+      >
+        <b-icon icon="chevron-down" />
+      </button>
+      <button
+        v-if="tableConfig.closedRows.has(rowData[tableConfig.rowId])"
+        class="btn btn-clear text-primary p-0 shadow-none"
+        @click="toggleRowDisplay(rowData[tableConfig.rowId])"
+      >
+        <b-icon icon="chevron-right" />
+      </button>
     </div>
+  </div>
 </template>
 
 <style lang="scss" scoped>
@@ -30,7 +37,10 @@
 </style>
 
 <script>
-    import datatableService from "./datatableService";
+import datatableService from './datatableService';
+// TODO fix eslint
+
+/* eslint-disable */
 
     export default {
         props: ['columnConfig', 'tableConfig', 'columnId', 'rowData'],
