+++ src/main.js
@@ -1,8 +1,2 @@
-import Vue from 'vue';
-import App from './App.vue';
-
-Vue.config.productionTip = false;
-
-new Vue({
-  render: (h) => h(App),
-}).$mount('#app');
+export { default as DatatableExample } from './Datatable/DatatableExample.vue';
+export default {};
