+++ src/stories/Datatable/Datatable.stories.js
@@ -1,0 +1,10 @@
+
+import { DatatableExample } from '../../main';
+
+export default { title: 'Deprecated Datatable' };
+
+
+export const Datatable = () => ({
+  components: { DatatableExample },
+  template: '<DatatableExample :rounded="true">rounded</DatatableExample>',
+});
