+++ src/stories/index.stories.js
@@ -1,1 +1,1 @@
-
+export * from './Datatable/Datatable.stories';
